fix: 2nd iteration of andoroid implementation
