+++ packages/react-native-audio-api/android/src/main/AndroidManifest.xml
@@ -1,6 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           package="com.swmansion.audioapi">
-  <application>
-    <activity android:name=".system.RecordingPermissionActivity" />
-  </application>
 </manifest>
