+++ packages/react-native-audio-api/android/src/main/AndroidManifestNew.xml
@@ -1,5 +1,2 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
-  <application>
-    <activity android:name=".system.RecordingPermissionActivity" />
-  </application>
 </manifest>
